refactor checklist section
