+++ features/page_objects/sections/checkbox_fee_section.rb
@@ -1,27 +1,5 @@
-class CheckboxFeeSection < SitePrism::Section
-  sections :checklist, '.multiple-choice' do
-    element :label, 'label'
-    element :checkbox, 'input[type="checkbox"]', visible: false
-  end
-
-  def checklist_labels
-    checklist.map { |item| item.label.text if item.has_label? }
-  end
-
-  def checklist_item_for(label)
-    checklist.find do |item|
-      item.label.text.match?(Regexp.new(Regexp.escape(label), true))
-    end
-  end
-
-  def check(label)
-    page.check checklist_item_for(label).checkbox['id'], visible: false
-  end
-
-  def uncheck(label)
-    page.uncheck checklist_item_for(label).checkbox['id'], visible: false
-  end
-
+require_relative 'checklist_section'
+class CheckboxFeeSection < ChecklistSection
   def set_quantity(label, value = 1)
     fee_block = fee_block_for(label)
     fee_block.quantity.set value
